Refactor decryption results into dedicated class
