+++ src/electionguard_cli/cli_models/__init__.py
@@ -1,7 +1,9 @@
-from electionguard_cli.cli_models.build_election_results import BuildElectionResults
+from electionguard_cli.cli_models.e2e_build_election_results import BuildElectionResults
 from electionguard_cli.cli_models.e2e_inputs import E2eInputs
+from electionguard_cli.cli_models.e2e_decrypt_results import E2eDecryptResults
 
 __all__ = [
     "BuildElectionResults",
-    "E2eInputs"
+    "E2eInputs",
+    "E2eDecryptResults"
 ]
