+++ src/electionguard_cli/commands/e2e_command.py
@@ -32,9 +32,9 @@ def e2e(guardian_count: int, quorum: int) -> None:
     joint_key = KeyCeremonyStep().run_key_ceremony(election_inputs)
     build_election_results = ElectionBuilderStep().build_election(election_inputs, joint_key)
     ballot_store = SubmitVotesStep().submit_votes(election_inputs, build_election_results)
-    (tally, spoiled_ballots) = DecryptStep().decrypt_tally(
+    decrypt_results = DecryptStep().decrypt_tally(
         ballot_store, election_inputs.guardians, build_election_results
     )
 
     # print results
-    PrintResultsStep().print_election_results(tally, spoiled_ballots)
+    PrintResultsStep().print_election_results(decrypt_results)
